+++ react-app/src/api/axios.tsx
@@ -11,9 +11,7 @@ import {
   IMember,
   IUsersAll,
   IPutUserProfile,
-} from "./interfaces";
-import { ReactNode, useContext, useEffect, useState } from "react";
-import { NotificationContext } from "./routes/NotificationProvider";
+} from "../interfaces";
 
 // const BASE_URL = "/api/";
 const BASE_URL = import.meta.env.VITE_SITE_URL + "/api/";
@@ -23,70 +21,6 @@ export const authApi = axios.create({
   baseURL: BASE_URL,
   withCredentials: true,
 });
-
-// authApi.interceptors.response.use(
-//   function (response) {
-//     // Any status code that lie within the range of 2xx cause this function to trigger
-//     // Do something with response data
-//     return response;
-//   },
-//   function (error) {
-//     // const foo = useAuth();
-
-//     // Any status codes that falls outside the range of 2xx cause this function to trigger
-//     // Do something with response error
-//     if (error.response.status) {
-//       // console.log(401, "error found baby!");
-//       // user?.logOut();
-//       // foo?.logOut()
-//       // console.log(AuthContext);
-//       console.log("Error intercepted:", error.response.status);
-//     }
-//     return Promise.reject(error);
-//     // https://stackoverflow.com/questions/62888255/how-to-use-react-usecontext-in-a-function-that-does-not-render-any-components
-//   },
-// );
-
-export function AxiosInterceptors({ children }: { children: ReactNode }) {
-  const [isSet, setIsSet] = useState(false);
-  const { addNotif } = useContext(NotificationContext);
-  useEffect(() => {
-    setIsSet(true);
-    const interceptor = authApi.interceptors.response.use(
-      (response) => {
-        return response;
-      },
-      (error) => {
-        if (error.response.status) {
-          addNotif({
-            from: error.response.status,
-            message: error.response.data.message,
-            type: "ERROR",
-          });
-        }
-
-        return Promise.reject(error);
-      },
-    );
-
-    return () => authApi.interceptors.response.eject(interceptor);
-  }, []);
-  console.log("isSet:          ", isSet);
-  return isSet && children;
-}
-
-// function AxiosIntercetor({children}) {
-//   const [isSet, setIsSet] = useState(false)
-
-//   useEffect(()=>{
-//     ...
-//     setIsSet(true)
-//   })
-
-//   return isSet && children
-// }
-
-// https://dev.to/arianhamdi/react-hooks-in-axios-interceptors-3e1h
 
 //-------------------------------------------User-------------------------------------------------------//
