+++ react-app/src/routes/ChatroomBrowser.tsx
@@ -9,7 +9,7 @@ import { LoadingSpinnerMessage } from "../components/Loading";
 import {
   useChatroomList,
   useMutPostNewChatroom,
-} from "../functions/customHook";
+} from "../api/apiHooks";
 import { IChatroomPost, IChatroom } from "../interfaces";
 import { Link, useNavigate } from "react-router-dom";
 import { useContext, useState } from "react";
